This is always the last 4 digits.
This was originally written to use everything but the first 3 digits.
Which worked for a long while!  That is, until the task IDs grew into a new
order of magnitude and it stopped working.

fedimg has been trying to curl this:

https://kojipkgs.fedoraproject.org//work/tasks/17094/10617094/Fedora-Cloud-Base-23_Alpha-20150805.x86_64.raw.xz

when it should have been trying to curl this:

https://kojipkgs.fedoraproject.org//work/tasks/7094/10617094/Fedora-Cloud-Base-23_Alpha-20150805.x86_64.raw.xz