Rearrange image.test fedmsg alerts.
First, tell the world that we are starting our test *before* we try to
deploy a node with our new AMI (in case that fails like it did today).

Second, if the deploy fails, then publish a second fedmsg message
indicating that our test fails (since we never even got to start the
"test" -- the node never came up).