Use new-style of accessing ec2 drivers.
According to the libcloud docs, the way you're supposed to access these
provider/driver objects changed in 0.14.0 and this is the way you're
supposed to do it now.  I tested this method by hand and was able to
create and destroy a temporary keypair in the eu-central-1 region, so..
I think it should work.

Please eyeball for typos as I haven't been able to test the full fedimg
cycle yet.

See https://libcloud.readthedocs.org/en/latest/upgrade_notes.html#amazon-ec2-compute-driver-changes