Pungi4 fixes.
With pungi4 we changed the way images are being created in koji.
They're no longer scratch tasks .. they're instead attached to a real
'build'.

https://fedorahosted.org/rel-eng/ticket/6304

This updates the code to handle both scenarios (images as scratch tasks
and images as real builds).  It seems we're still producing a small
handful of them the old way.